# Conflicts:
#	connectors/katalog/pkg/connector/credentials.go
#	go.mod
#	go.sum
#	manager/apis/app/v1alpha1/v1alpha1_suite_test.go
#	manager/config/crd/bases/app.m4d.ibm.com_blueprints.yaml
#	manager/config/crd/bases/app.m4d.ibm.com_m4dapplications.yaml
#	manager/config/crd/bases/app.m4d.ibm.com_m4dmodules.yaml
#	manager/config/crd/bases/app.m4d.ibm.com_m4dstorageaccounts.yaml
#	manager/config/crd/bases/app.m4d.ibm.com_plotters.yaml
#	manager/config/crd/bases/motion.m4d.ibm.com_batchtransfers.yaml
#	manager/config/crd/bases/motion.m4d.ibm.com_streamtransfers.yaml
#	manager/config/webhook/manifests.yaml
#	manager/controllers/app/blueprint_controller_test.go
#	manager/controllers/app/m4dapplication_controller_test.go
#	manager/controllers/motion/suite_test.go

+++ hack/tools/create_kind.sh
@@ -10,14 +10,17 @@ source ./common.sh
 
 K8S_VERSION=${K8S_VERSION:-v1.19.1}
 : ${KUBE_NAMESPACE:=m4d-system}
+: ${KEEP_PROXY:=true}
 
 registry_delete() {
   docker network disconnect kind kind-registry
   docker kill kind-registry
   docker rm -f kind-registry
-  docker network disconnect kind kind-registry-proxy
-  docker kill kind-registry-proxy
-  docker rm -f kind-registry-proxy
+  if [ "${KEEP_PROXY}" != true ]; then
+    docker network disconnect kind kind-registry-proxy
+    docker kill kind-registry-proxy
+    docker rm -f kind-registry-proxy
+  fi
 }
 
 registry_create() {
